extra Rate Limit buffering (if user says they're using meow.camera)
I refactored the code to include a check for the usage of meow.camera.
If meow.camera is being used,
there is an extra rate limit buffering mechanism in place to wait for the rate limit to reset when there are only 4 requests remaining.
This should hopefully make meow.camera (and other sites that use meow.camera's API directly) more usable.